+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/generate_code/CodeGeneratorTest.java
@@ -57,42 +57,42 @@ class CodeGeneratorTest {
 
 	@BeforeEach
 	void setUp() throws Exception {
-		this.context = new AnnotationConfigApplicationContext(AllGraphQLCases_Server_SpringConfiguration.class);
-		this.pluginConfiguration = this.context.getBean(GraphQLConfigurationTestHelper.class);
-		this.mavenTestHelper = this.context.getBean(MavenTestHelper.class);
+		context = new AnnotationConfigApplicationContext(AllGraphQLCases_Server_SpringConfiguration.class);
+		pluginConfiguration = context.getBean(GraphQLConfigurationTestHelper.class);
+		mavenTestHelper = context.getBean(MavenTestHelper.class);
 
-		this.targetResourceFolder = this.mavenTestHelper.getTargetResourceFolder(this.getClass().getSimpleName());
-		this.targetSourceFolder = this.mavenTestHelper.getTargetSourceFolder(this.getClass().getSimpleName());
-		this.targetRuntimeClassesSourceFolder = this.mavenTestHelper
+		targetResourceFolder = mavenTestHelper.getTargetResourceFolder(this.getClass().getSimpleName());
+		targetSourceFolder = mavenTestHelper.getTargetSourceFolder(this.getClass().getSimpleName());
+		targetRuntimeClassesSourceFolder = mavenTestHelper
 				.getTargetRuntimeClassesBaseSourceFolder(this.getClass().getSimpleName());
-		this.testRuntimeSourcesFile = this.mavenTestHelper.getTestRutimeSourcesJarFile();
+		testRuntimeSourcesFile = mavenTestHelper.getTestRutimeSourcesJarFile();
 
-		if (this.targetResourceFolder.exists()) {
-			FileUtils.forceDelete(this.targetResourceFolder);
-			this.targetResourceFolder.mkdirs();
+		if (targetResourceFolder.exists()) {
+			FileUtils.forceDelete(targetResourceFolder);
+			targetResourceFolder.mkdirs();
 		}
-		if (this.targetSourceFolder.exists()) {
-			FileUtils.forceDelete(this.targetSourceFolder);
-			this.targetSourceFolder.mkdirs();
-		}
-
-		if (this.testRuntimeSourcesFile.exists()) {
-			this.testRuntimeSourcesFile.delete();
+		if (targetSourceFolder.exists()) {
+			FileUtils.forceDelete(targetSourceFolder);
+			targetSourceFolder.mkdirs();
 		}
 
-		this.codeGenerator = this.context.getBean(GenerateCodeGenerator.class);
-		this.codeGenerator.generateCodeDocumentParser = new GenerateCodeDocumentParser(this.pluginConfiguration);
+		if (testRuntimeSourcesFile.exists()) {
+			testRuntimeSourcesFile.delete();
+		}
+
+		codeGenerator = context.getBean(GenerateCodeGenerator.class);
+		codeGenerator.generateCodeDocumentParser = new GenerateCodeDocumentParser(pluginConfiguration);
 	}
 
 	@AfterEach
 	void close() {
-		this.context.close();
+		context.close();
 	}
 
 	@Test
 	@Execution(ExecutionMode.CONCURRENT)
 	void testCodeGenerator() {
-		assertNotNull(this.codeGenerator.velocityEngine, "Velocity engine must be initialized");
+		assertNotNull(codeGenerator.velocityEngine, "Velocity engine must be initialized");
 	}
 
 	/**
@@ -104,14 +104,14 @@ class CodeGeneratorTest {
 	@Execution(ExecutionMode.CONCURRENT)
 	void test_generateTargetFile_client() {
 		// Let's mock the Velocity engine, to check how it is called
-		this.codeGenerator.velocityEngine = mock(VelocityEngine.class);
+		codeGenerator.velocityEngine = mock(VelocityEngine.class);
 		Template mockedTemplate = mock(Template.class);
-		when(this.codeGenerator.velocityEngine.getTemplate(anyString(), anyString())).thenReturn(mockedTemplate);
+		when(codeGenerator.velocityEngine.getTemplate(anyString(), anyString())).thenReturn(mockedTemplate);
 
-		this.codeGenerator.generateCodeDocumentParser = mock(GenerateCodeDocumentParser.class);
-		this.pluginConfiguration.mode = PluginMode.client;
+		codeGenerator.generateCodeDocumentParser = mock(GenerateCodeDocumentParser.class);
+		pluginConfiguration.mode = PluginMode.client;
 
-		ObjectType object1 = new ObjectType("O1", this.pluginConfiguration, null);
+		ObjectType object1 = new ObjectType("O1", pluginConfiguration, null);
 		// ObjectType object2 = new ObjectType("O2", configuration);
 		List<Type> objects = new ArrayList<>();
 		objects.add(object1);
@@ -120,7 +120,7 @@ class CodeGeneratorTest {
 		TargetFileType type = TargetFileType.OBJECT;
 
 		// Go, go, go
-		int i = this.codeGenerator.generateTargetFilesForTypeList(objects, type, CodeTemplate.OBJECT, false);
+		int i = codeGenerator.generateTargetFilesForTypeList(objects, type, CodeTemplate.OBJECT, false);
 
 		// Verification
 		assertEquals(objects.size(), i, "Nb files generated");
@@ -128,7 +128,7 @@ class CodeGeneratorTest {
 		// Let's check the parameter for getTemplate
 		ArgumentCaptor<String> argument1 = ArgumentCaptor.forClass(String.class);
 		ArgumentCaptor<String> argument2 = ArgumentCaptor.forClass(String.class);
-		verify(this.codeGenerator.velocityEngine, times(1)).getTemplate(argument1.capture(), argument2.capture());
+		verify(codeGenerator.velocityEngine, times(1)).getTemplate(argument1.capture(), argument2.capture());
 		assertEquals(CodeTemplate.OBJECT.getDefaultPath(), argument1.getValue(),
 				"checks the parameter for getTemplate");
 		assertEquals("UTF-8", argument2.getValue());
@@ -139,7 +139,7 @@ class CodeGeneratorTest {
 		ArgumentCaptor<Context> argumentContext = ArgumentCaptor.forClass(Context.class);
 		verify(mockedTemplate, times(1)).merge(argumentContext.capture(), any(Writer.class));
 		// We have the Context sent to the Template.merge(..) method. Let's check its content
-		assertEquals(this.pluginConfiguration.getPackageName(),
+		assertEquals(pluginConfiguration.getPackageName(),
 				((GraphQLConfiguration) argumentContext.getValue().get("configuration")).getPackageName(),
 				"Context: checks the package");
 		assertEquals(object1, argumentContext.getValue().get("object"), "Context: checks the package");
@@ -156,20 +156,20 @@ class CodeGeneratorTest {
 	@Execution(ExecutionMode.CONCURRENT)
 	void test_generateTargetFile_server() {
 		// Let's mock the Velocity engine, to check how it is called
-		this.codeGenerator.velocityEngine = mock(VelocityEngine.class);
+		codeGenerator.velocityEngine = mock(VelocityEngine.class);
 		Template mockedTemplate = mock(Template.class);
-		when(this.codeGenerator.velocityEngine.getTemplate(anyString(), anyString())).thenReturn(mockedTemplate);
+		when(codeGenerator.velocityEngine.getTemplate(anyString(), anyString())).thenReturn(mockedTemplate);
 
-		this.pluginConfiguration.mode = PluginMode.server;
+		pluginConfiguration.mode = PluginMode.server;
 
-		ObjectType object1 = new ObjectType("O1", this.pluginConfiguration, null);
+		ObjectType object1 = new ObjectType("O1", pluginConfiguration, null);
 		List<Type> objects = new ArrayList<>();
 		objects.add(object1);
 
 		TargetFileType type = TargetFileType.OBJECT;
 
 		// Go, go, go
-		int i = this.codeGenerator.generateTargetFilesForTypeList(objects, type, CodeTemplate.OBJECT, false);
+		int i = codeGenerator.generateTargetFilesForTypeList(objects, type, CodeTemplate.OBJECT, false);
 
 		// Verification
 		assertEquals(objects.size(), i, "Nb files generated");
@@ -177,7 +177,7 @@ class CodeGeneratorTest {
 		// Let's check the parameter for getTemplate
 		ArgumentCaptor<String> argument1 = ArgumentCaptor.forClass(String.class);
 		ArgumentCaptor<String> argument2 = ArgumentCaptor.forClass(String.class);
-		verify(this.codeGenerator.velocityEngine, times(1)).getTemplate(argument1.capture(), argument2.capture());
+		verify(codeGenerator.velocityEngine, times(1)).getTemplate(argument1.capture(), argument2.capture());
 		assertEquals(CodeTemplate.OBJECT.getDefaultPath(), argument1.getValue(),
 				"checks the parameter for getTemplate");
 		assertEquals("UTF-8", argument2.getValue());
@@ -188,7 +188,7 @@ class CodeGeneratorTest {
 		ArgumentCaptor<Context> argumentContext = ArgumentCaptor.forClass(Context.class);
 		verify(mockedTemplate, times(1)).merge(argumentContext.capture(), any(Writer.class));
 		// We have the Context sent to the Template.merge(..) method. Let's check its content
-		assertEquals(this.pluginConfiguration.getPackageName(),
+		assertEquals(pluginConfiguration.getPackageName(),
 				((GraphQLConfiguration) argumentContext.getValue().get("configuration")).getPackageName(),
 				"Context: checks the package");
 		assertEquals(object1, argumentContext.getValue().get("object"), "Context: checks the package");
@@ -201,10 +201,11 @@ class CodeGeneratorTest {
 	void test_generateOneFile_inClasspath() throws IOException {
 		// Preparation
 		VelocityContext velocityContext = new VelocityContext();
-		File targetFile = new File(this.targetResourceFolder + "/testTemplate_inClasspath");
+		velocityContext.put("objectContentTemplate", codeGenerator.resolveTemplate(CodeTemplate.OBJECT_CONTENT));
+		File targetFile = new File(targetResourceFolder + "/testTemplate_inClasspath");
 
 		// Go, go, go
-		this.codeGenerator.generateOneFile(targetFile, "In test_generateOneFile_InClasspath", velocityContext,
+		codeGenerator.generateOneFile(targetFile, "In test_generateOneFile_InClasspath", velocityContext,
 				CodeTemplate.OBJECT);
 
 		// If there is no error, then the template has been found. The test is Ok
@@ -216,46 +217,46 @@ class CodeGeneratorTest {
 		// Preparation
 		String name = "MyClass";
 		String packageName = "my.package";
-		this.pluginConfiguration.packageName = packageName;
-		this.pluginConfiguration.separateUtilityClasses = false;
-		this.pluginConfiguration.targetResourceFolder = this.targetResourceFolder;
-		this.pluginConfiguration.targetSourceFolder = this.targetSourceFolder;
+		pluginConfiguration.packageName = packageName;
+		pluginConfiguration.separateUtilityClasses = false;
+		pluginConfiguration.targetResourceFolder = targetResourceFolder;
+		pluginConfiguration.targetSourceFolder = targetSourceFolder;
 
 		////////////////////////////////////////////////////////////////////////////////////////////////////
 		// Go, go, go (separateUtilityClasses=false, isUtility=false)
-		File file = this.codeGenerator.getJavaFile(name, false);
+		File file = codeGenerator.getJavaFile(name, false);
 		// Verification
-		String expectedEndOfPath = (this.targetSourceFolder.getCanonicalPath() + '/' + packageName + '/' + name)
+		String expectedEndOfPath = (targetSourceFolder.getCanonicalPath() + '/' + packageName + '/' + name)
 				.replace('.', '/').replace('\\', '/') + ".java";
 		assertEquals(expectedEndOfPath, file.getCanonicalPath().replace('\\', '/'), "The file path should end with "
 				+ expectedEndOfPath + ", but is " + file.getCanonicalPath().replace('\\', '/'));
 
 		////////////////////////////////////////////////////////////////////////////////////////////////////
 		// Go, go, go (separateUtilityClasses=false, isUtility=true)
-		file = this.codeGenerator.getJavaFile(name, true);
+		file = codeGenerator.getJavaFile(name, true);
 		// Verification
-		expectedEndOfPath = (this.targetSourceFolder.getCanonicalPath() + '/' + packageName + '/' + name)
-				.replace('.', '/').replace('\\', '/') + ".java";
+		expectedEndOfPath = (targetSourceFolder.getCanonicalPath() + '/' + packageName + '/' + name).replace('.', '/')
+				.replace('\\', '/') + ".java";
 		assertEquals(expectedEndOfPath, file.getCanonicalPath().replace('\\', '/'), "The file path should end with "
 				+ expectedEndOfPath + ", but is " + file.getCanonicalPath().replace('\\', '/'));
 
 		// separateUtilityClasses=true
-		this.pluginConfiguration.separateUtilityClasses = true;
+		pluginConfiguration.separateUtilityClasses = true;
 
 		////////////////////////////////////////////////////////////////////////////////////////////////////
 		// Go, go, go (separateUtilityClasses=true, isUtility=false)
-		file = this.codeGenerator.getJavaFile(name, false);
+		file = codeGenerator.getJavaFile(name, false);
 		// Verification
-		expectedEndOfPath = (this.targetSourceFolder.getCanonicalPath() + '/' + packageName + '/' + name)
-				.replace('.', '/').replace('\\', '/') + ".java";
+		expectedEndOfPath = (targetSourceFolder.getCanonicalPath() + '/' + packageName + '/' + name).replace('.', '/')
+				.replace('\\', '/') + ".java";
 		assertEquals(expectedEndOfPath, file.getCanonicalPath().replace('\\', '/'), "The file path should end with "
 				+ expectedEndOfPath + ", but is " + file.getCanonicalPath().replace('\\', '/'));
 
 		////////////////////////////////////////////////////////////////////////////////////////////////////
 		// Go, go, go (separateUtilityClasses=true, isUtility=true)
-		file = this.codeGenerator.getJavaFile(name, true);
+		file = codeGenerator.getJavaFile(name, true);
 		// Verification
-		expectedEndOfPath = (this.targetSourceFolder.getCanonicalPath() + '/' + packageName + '/'
+		expectedEndOfPath = (targetSourceFolder.getCanonicalPath() + '/' + packageName + '/'
 				+ GenerateCodeDocumentParser.UTIL_PACKAGE_NAME + '/' + name).replace('.', '/').replace('\\', '/')
 				+ ".java";
 		assertEquals(expectedEndOfPath, file.getCanonicalPath().replace('\\', '/'), "The file path should end with "
@@ -272,16 +273,16 @@ class CodeGeneratorTest {
 	@Execution(ExecutionMode.CONCURRENT)
 	void testGenerateCode_noCopyOfRuntimeSources() throws IOException {
 
-		this.pluginConfiguration.mode = PluginMode.client;
-		this.pluginConfiguration.packageName = "test.generatecode.enabled";
-		this.pluginConfiguration.schemaFileFolder = new File("src/test/resources");
-		this.pluginConfiguration.schemaFilePattern = "basic.graphqls";
-		this.pluginConfiguration.targetResourceFolder = this.targetResourceFolder;
-		this.pluginConfiguration.targetSourceFolder = this.targetSourceFolder;
-		this.pluginConfiguration.targetClassFolder = this.targetSourceFolder;
+		pluginConfiguration.mode = PluginMode.client;
+		pluginConfiguration.packageName = "test.generatecode.enabled";
+		pluginConfiguration.schemaFileFolder = new File("src/test/resources");
+		pluginConfiguration.schemaFilePattern = "basic.graphqls";
+		pluginConfiguration.targetResourceFolder = targetResourceFolder;
+		pluginConfiguration.targetSourceFolder = targetSourceFolder;
+		pluginConfiguration.targetClassFolder = targetSourceFolder;
 
-		this.codeGenerator.generateCode();
-		assertFalse(this.targetRuntimeClassesSourceFolder.exists());
+		codeGenerator.generateCode();
+		assertFalse(targetRuntimeClassesSourceFolder.exists());
 	}
 
 	/**
@@ -290,8 +291,8 @@ class CodeGeneratorTest {
 	@Test
 	@Execution(ExecutionMode.CONCURRENT)
 	protected void testResolveTemplateDefault() {
-		this.pluginConfiguration.templates.clear();
-		assertEquals(CodeTemplate.WIRING.getDefaultPath(), this.codeGenerator.resolveTemplate(CodeTemplate.WIRING));
+		pluginConfiguration.templates.clear();
+		assertEquals(CodeTemplate.WIRING.getDefaultPath(), codeGenerator.resolveTemplate(CodeTemplate.WIRING));
 		;
 	}
 
@@ -301,9 +302,9 @@ class CodeGeneratorTest {
 	@Test
 	@Execution(ExecutionMode.CONCURRENT)
 	protected void testResolveTemplateCustom() {
-		this.pluginConfiguration.templates.clear();
-		this.pluginConfiguration.templates.put(CodeTemplate.WIRING.name(), "/my/custom/template");
-		assertEquals("/my/custom/template", this.codeGenerator.resolveTemplate(CodeTemplate.WIRING));
+		pluginConfiguration.templates.clear();
+		pluginConfiguration.templates.put(CodeTemplate.WIRING.name(), "/my/custom/template");
+		assertEquals("/my/custom/template", codeGenerator.resolveTemplate(CodeTemplate.WIRING));
 		;
 	}
 
@@ -314,7 +315,7 @@ class CodeGeneratorTest {
 	 * @throws IOException
 	 */
 	protected void createRuntimeSourcesJar() throws IOException {
-		File file = this.mavenTestHelper.getTestRutimeSourcesJarFile();
+		File file = mavenTestHelper.getTestRutimeSourcesJarFile();
 		if (file.exists()) {
 			file.delete();
 		}